Prototype integration into save/load and simple correctness test
There are many things this doesn't handle properly
- byteorder
- map_location etc.
- ...

These are all kinks we can smooth out, this is intended as a simple
example for now.

There is a script save_load_gds_sample.py one can run to verify that checkpoints
saved with GDS can be loaded normally and vice versa. Note that the
filesystem must be ext4, see my earlier commit Upstream gdsfile from apex without api changes
forr a description of how to mount and ext4 filesystem on a loop device